Test QC_proc and modification parent_clustering
I tested the QC_proc on the different dataset to be sure that the updated procedure is functionning well. I also implemented haplotypes graph production in parent_cluster.

QC_proc():

1.	Problem with cross.ind after removing too small crosses

Reorganised the order of the different operation performed in QC_proc and ensure that argument dependent on the list of marker and genotype are progressively updated within the function.

2.	For the ABH assignment taking sub-cross structure into consideration

I added to extra arguments for QC_proc() subcross.ind and par.per.subcross. The subcross appartenance of each genotypes can be specified in the subcross.ind vector and par.per.subcross is similar to par.per.cross.

3.	I also added a modification to QC_proc to keep more markers. In the previous version, any marker that showed a problematic within cross MAF in at least one cross was removed. Now, using argument MAF.cr.miss = TRUE, it is possible to set the problematic missing within the cross where they show too little segregation.

4.	For position with more than one marker, I select now the one that segregate the most.

5.	QC_proc also return now par.per.cross that can be used in the next functions. I attached an updated script of the workflow.

Parent_cluster():

I added two extra argument to produce the haplotypes plots (plot = TRUE) and save them in a specified location (plot.loc = TRUE).

Concerning the warnings retunred by clusthaplo, I did not have a very close look to it but I do not really know how to silence these warnings. It seems that there is no direct argument in the function that could do that.