Modification of the CV procedure
This modification consist in small updates in the way pTS and pVS are defined. I also reduced the amount of results that are give back to the user

pTS is R2.adj/her where her is the average within cross heritability.

pVS is computed within cross using the squared Pearson correlation between the observed and predicted values. It is corrected by the heritability. The user can provide a vector given the within cross heritabilities.

Two value for pVS are returned the within cross pVS in a matrix (p.vs.cr) and a global p.vs doing the average within cross p.vs.

For the individual QTL CV statistics I give the adjusted QTL partial R2 by difference in TS and QTL partial R2 in VS plus the bias.