order parental/ancestral allele MAF
In the QTL incidence matrix of the parental and ancestral model, the allele are automatically ordered from the less segregating (loweset MAF) to the most (highest MAF) from left to the right.

The reason for that is to make that in any case the reference allele will be the most segregating one. This is the most consistent way to set the constraint this is also practically a good solution.

This modification applied in SIM, CIM, MQE_CIM profiles computation also in backward elimination functions and R2 computation.