Modification estimation QTL gen effects
The biggest modification of this commit concern s the estimation of the genetic effects. The constraint for the parental and ancestral model are now set as the most frequent allele within each connected part. Therfore, I created the sub-function: IncMat_QTL_MAF to re-order the allele.

I modify the function QTL_genEffects and MQE_genEffects and especially the sub-function used to organise the results (Qeff_res_processing and Qeff_res_processing_MQE)

I also modify the function and documentation of the function using QTL_genEffects and MQE_genEffects. for example mpp_proc or MQE_proc.

I also made few modification following the modification done for the TAG paper review. In the CV procedure I compute pVS as a weighted mean. I also extended the checks that are done in the MQE procedure.

Finally I solved two bug detected in KWS analysis. One concern the QC_proc and the presence of maker that are fully monomorphic in 1 cross. Now these marker do not make problem when I want to replace their value with missing.

In the CV procedure there was also some problem when the estimation can not be done for one cross due to not complete case information. This is especially problematic for the bi-allelic model. We need to use more imputed data for the bi-allelic model