Parent reference estimation QTL effects
Addition of a possibility for the user to specify a parent reference for the estimation of the QTL effect. I added the argument ref.par in the QTL_genEffects() function. This option is only available if there is only one connected part in the MPP design.

The mppR version is updated to 1.0.1