Modification of QC_proc for imputation of IBS data
I included an option to impute the IBS data after the QC. This option call the codeGeno() function from synbreed. The call of Beagle look a bit unstable (at least on my computer). In some cases, it works in other it seems that there are some problems to make the connection with Beagle.