Modification bi-allelic QTL effects report
So far I gave to AA additive effect estimated for the bi-allelic model. This effect represent in fact the effect of a single minor allele copy. Therefore AT should get this effect and AA should get 2* the additive effect.

Both AA and AT can get the same significance for the effect.