modification mppData_subset
So far in mppData_subset, the geno.par argument was not modified if the number of parents of the population was reduced. This created some problem for the estimation of the genetic effects of the bi-allelic model.

In that modification, the parent column that are not anymore used in the reduced data are removed.