create QC.mppData
create a new function to perform the quality control on raw data. This function inherit from QC_proc the part dedicated to the QC.

The part dedicated to the ABH transformation and the imputation will be transfered to other functions (IBS.mppData and IBD.mppData).

All internal QC function will be silenced.

Creation of a new example data object mppData_init. which is an initial data object with the raw data before any processing.