modification compute QTL model without par.clu for cr, par and biall models
The QTL detection model can be computed for the cross-specific, parental and bi-allelic model if the mppData object is not completed with the par.clu object. So the par.clu object is an option for the ancestral model.

For the ancestral model check that the par.clu object is there.